requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.2
python-multipart==0.0.12
anthropic==0.40.0
python-dotenv==1.0.1
httpx==0.27.2
